v3-1
Код обработки изображенй - фото не на диск, как было в пред версиях, а в озу
